// ClientApp/src/app/models/eco-info.ts
export class EcoInfo {
    DASH_8: string;
    AD_IND: string;
    WORK_REQUIREMENT: string;
    STATION_REQUIREMENTS: string;
    REVISION_DATE: Date;
    EO_NUMBER: string;
    OPERATION_STEP: number;
    TASK_TYPE: string;
    ME_TYPE: string;
    MAINT_TYPE: string;
    COMMENTS: string;
    SCHEDULING_STATUS: string;
    DATA_SOURCE: string;
    ATA: string;
    MFG_PART_NUMBER: string;
    EO_DESC: string;
    CNT_SCHDLD: number;
    NO_OF_ACCOMPLISHMENTS: number;
    DASH8_COMMENTS: string;
    EO_HISTORY: string;
    UPDATED_DATETIME: Date;
}